+++ src/write_delim.cpp
@@ -1,9 +1,12 @@
-#include <Rcpp.h>
-using namespace Rcpp;
+#include "cpp11/list.hpp"
+#include "cpp11/sexp.hpp"
+#include "cpp11/strings.hpp"
+
 #include "grisu3.h"
 #include "write_connection.h"
 #include <boost/iostreams/stream.hpp> // stream
 #include <fstream>
+#include <sstream>
 
 enum quote_escape_t { DOUBLE = 1, BACKSLASH = 2, NONE = 3 };
 
@@ -11,7 +14,7 @@ enum quote_escape_t { DOUBLE = 1, BACKSLASH = 2, NONE = 3 };
 template <class Stream>
 void stream_delim(
     Stream& output,
-    const RObject& x,
+    const cpp11::sexp& x,
     int i,
     char delim,
     const std::string& na,
@@ -20,7 +23,7 @@ void stream_delim(
 template <class Stream>
 void stream_delim_row(
     Stream& output,
-    const Rcpp::List& x,
+    const cpp11::list& x,
     int i,
     char delim,
     const std::string& na,
@@ -88,7 +91,7 @@ void stream_delim(
 template <class Stream>
 void stream_delim(
     Stream& output,
-    const List& df,
+    const cpp11::list& df,
     char delim,
     const std::string& na,
     bool col_names,
@@ -104,7 +107,7 @@ void stream_delim(
   }
 
   if (col_names) {
-    CharacterVector names = as<CharacterVector>(df.attr("names"));
+    cpp11::strings names(df.attr("names"));
     for (int j = 0; j < p; ++j) {
       stream_delim(output, names, j, delim, na, escape);
       if (j != p - 1)
@@ -113,7 +116,7 @@ void stream_delim(
     output << eol;
   }
 
-  RObject first_col = df[0];
+  cpp11::sexp first_col = df[0];
   int n = Rf_length(first_col);
 
   for (int i = 0; i < n; ++i) {
@@ -121,10 +124,9 @@ void stream_delim(
   }
 }
 
-// [[Rcpp::export]]
-std::string stream_delim_(
-    const List& df,
-    RObject connection,
+[[cpp11::register]] std::string stream_delim_(
+    const cpp11::list& df,
+    cpp11::sexp connection,
     char delim,
     const std::string& na,
     bool col_names,
@@ -167,7 +169,7 @@ std::string stream_delim_(
 template <class Stream>
 void stream_delim(
     Stream& output,
-    const RObject& x,
+    const cpp11::sexp& x,
     int i,
     char delim,
     const std::string& na,
@@ -220,7 +222,7 @@ void stream_delim(
     break;
   }
   default:
-    Rcpp::stop(
+    cpp11::stop(
         "Don't know how to handle vector of type %s.", Rf_type2char(TYPEOF(x)));
   }
 }