+++ tests/testthat.R
@@ -1,8 +1,4 @@
 library(testthat)
 library(readr)
 
-if (requireNamespace("xml2")) {
-  test_check("readr", reporter = MultiReporter$new(reporters = list(JunitReporter$new(file = "test-results.xml"), CheckReporter$new())))
-} else {
-  test_check("readr")
-}
+test_check("readr")