+++ tests/testthat/test-parsing-time.R
@@ -5,8 +5,8 @@ test_that("default format captures cases", {
   expect_equal(parse_time("22:20"), late_night)
   expect_equal(parse_time("10:20 pm"), late_night)
 
-  expect_equal(parse_time("22:20:05"), hms::as.hms(late_night + 5))
-  expect_equal(parse_time("10:20:05 pm"), hms::as.hms(late_night + 5))
+  expect_equal(parse_time("22:20:05"), hms::as_hms(late_night + 5))
+  expect_equal(parse_time("10:20:05 pm"), hms::as_hms(late_night + 5))
 })
 
 test_that("twelve o'clock is parsed properly", {
@@ -40,4 +40,15 @@ test_that("times are guessed as expected", {
 
   expect_equal(
     guess_parser(c("04:00:00", "04:30:00", "14:00:22")), "time")
+
+  expect_equal(
+    guess_parser("25:01:01"), "time")
+})
+
+test_that("durations", {
+  expect_warning(parse_time("25:00:00", format = "%H:%M:%S"))
+  expect_equal(parse_time("25:00:00", format = "%h:%M:%S"), hms::hms(hours = 25))
+  expect_equal(parse_time("1000000000:00:00", format = "%h:%M:%S"), hms::hms(hours = 1e9))
+  expect_equal(parse_time("-1:23:45", format = "%h:%M:%S"), hms::as_hms(-hms::hms(45, 23, 1)))
+  expect_equal(parse_time("-1:23:45.67", format = "%h:%M:%OS"), hms::as_hms(-hms::hms(45.67, 23, 1)))
 })