+++ tests/testthat/test-write.R
@@ -25,7 +25,8 @@ test_that("read_delim/csv/tsv and write_delim round trip special chars", {
   input <- read_delim(format_delim(output, delim = " "), delim = " ", trim_ws = FALSE, progress = FALSE)
   input_csv <- read_csv(format_delim(output, delim = ","), trim_ws = FALSE, progress = FALSE)
   input_tsv <- read_tsv(format_delim(output, delim = "\t"), trim_ws = FALSE, progress = FALSE)
-  expect_equal(input$x, input_csv$x, input_tsv$x,  x)
+  expect_equal(input$x, input_csv$x)
+  expect_equal(input_tsv$x,  x)
 })
 
 test_that("special floating point values translated to text", {
@@ -178,7 +179,7 @@ test_that("Can change the escape behavior for quotes", {
 })
 
 test_that("hms NAs are written without padding (#930)", {
-  df <- data.frame(x = hms::as.hms(c(NA, 34.234)))
+  df <- data.frame(x = hms::as_hms(c(NA, 34.234)))
   expect_equal(format_tsv(df), "x\nNA\n00:00:34.234\n")
 })