+++ NEWS.md
@@ -28,6 +28,8 @@
 * `guess_parser()` gains a `na` argument and removes NA values before guessing.
   `parse_guess()` now passes the `na` argument to `guess_parser()` (#1041).
 
+* New `%h` placeholder for parsing unrestricted hours (<0 and >23) to support parsing durations (#549, @krlmlr).
+
 * Uses of `tibble::data_frame` updated to `tibble::tibble`
 ([tidyverse/dplyr#4069](https://github.com/tidyverse/dplyr/issues/4069),
 @thays42)
@@ -57,7 +59,7 @@
 
 * `type_convert()` removes a 'spec' attribute, because the current columns likely have modified data types.  The 'spec' attribute is set by functions like `read_delim()` (@jimhester, @wibeasley, #1032).
 
-* `write_rds()` now can specify the rds version to use. The default value is
+* `write_rds()` now can specify the Rds version to use. The default value is
   2 as it's compatible to R versions prior to 3.5.0 (@shrektan, #1001).
 
 # readr 1.3.1